Add debounce prevention when led are set to 0 brightness
When the brightness gets to 0 keep it at that and ignore further rotary
encoder updates for up to 100ms. This is a workaround on a 'debounce'
effect on the rotary encoder, where it can sometimes read noise values
when turned too fast and bounce back to positive even though it's been
turning counter clockwise.